Rename package in jena-dboe-index-test
