+++ jena-db/jena-dboe-trans-data/src/test/java/org/apache/jena/dboe/trans/bplustree/TestBPlusTreeTxn.java
@@ -19,7 +19,7 @@
 package org.apache.jena.dboe.trans.bplustree;
 
 import org.apache.jena.dboe.base.file.Location;
-import org.apache.jena.dboe.index.test.IndexTestLib;
+import org.apache.jena.dboe.index.testlib.IndexTestLib;
 import org.apache.jena.system.Txn;
 import org.apache.jena.dboe.test.RecordLib;
 import org.apache.jena.dboe.transaction.Transactional;
